pyproject.toml
[tool.poetry]
name = "django_default_project"
version = "0.8.0"
description = "Its not always good to start from scratch .. lets have some sane defaults .."
authors = ["Andy Grabow <andy@freilandkiwis.de>"]
license = "MIT"
readme = 'README.md'

repository = "https://github.com/kakulukia/django-default-project"
homepage = "https://github.com/kakulukia/django-default-project"

[tool.poetry.dependencies]
django = "^4.0"
django-secrets = "^1.0.2"
django-undeletable = "1.1.2"
django-extensions = "^3.1.1"
dj-static = "^0.0.6"
django-compressor = "^4.1"
django-axes = "^5.2.0"
raven = "^6.10.0"
pypugjs = "^5.8.1"
django-post-office = "^3.2.1"
uwsgidecorators = "^1.1.0"
huepy = "^1.2.1"
djangorestframework = "^3.11.0"
requests = "^2.22.0"
pendulum = "^2.0.5"
Pillow = "^9.0.1"
python = "^3.9.9"
sentry-sdk = "^1.5.7"
django-cors-headers = "^3.13.0"

[tool.poetry.group.dev.dependencies]
black = "^22.1.0"
django-debug-toolbar = "^3.2"
django-livereload-server = "^0.4"
fabric = "^2.5.0"
ipdb = "^0.13.6"
djdt-flamegraph = "^0.2.12"
isort = "^5.10.1"
flake8 = "^5.0.4"

[tool.black]
line-length = 99
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
/migrations/  # exclude django migrations
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[flake8]
exclude = ['.git', 'migrations']
max-line-length = 140
